src/backend/coach-engine/requirements.txt
# FILE: src/coach-engine/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
